ml/dppi/requirements.txt
catboost>=1.2
numpy>=1.26
pandas>=2.2
scikit-learn>=1.4
pyyaml>=6.0
psycopg[binary]>=3.2
pyarrow>=16.0
joblib>=1.4
python-dotenv>=1.0
requests>=2.32
